==> Scripts/06_ensemble.R <==
# ====================================================================================================== #
# Description
#
#   Ensemble
#
# Change log:
#   Ver   Date        Comment
#   1.0   03/04/23    Initial version
#
# ====================================================================================================== #
# ------------------------------------------------------------------------------------------------------ #
# LIBRARIES
# ------------------------------------------------------------------------------------------------------ #

library(data.table)
library(tidyverse)
library(magrittr)

# ------------------------------------------------------------------------------------------------------ #
# IMPORT AND SOURCES
# ------------------------------------------------------------------------------------------------------ #

list_preds <- 
  list.files("./Output") %>% 
  grep("\\.csv", x = ., value = TRUE) %>% 
  paste0("./Output/", .) %>% 
  map(fread)
  

# ------------------------------------------------------------------------------------------------------ #
# PROGRAM
# ------------------------------------------------------------------------------------------------------ #

df_subm <- 
  bind_cols(
    list_preds
  ) %>% 
  mutate(
    id = id...1,
    cost = rowMeans(pick(starts_with("cost")))
  ) %>% 
  select(
    id,
    cost
  )
  

# ==== EXPORT ------------------------------------------------------------------------------------------ 

df_subm %>% 
  fwrite(file = "./Output/05_submission_ens.csv")
